requirements.txt
django 
djangorestframework 
django-environ 
pillow 
drf-yasg 
django-cors-headers